Get fmuzip.c to compile
Because the APIs for getting environment variables as well as getting and setting the current directory are so different between Windows and Linux, I just created two different versions of the fmuUnzip function.  It may be that by using helper routines and/or macros I can minimize the redundancy, but this works for now.